- Small changes and fixes

+++ wp-gdpr-compliance/uninstall.php
@@ -7,8 +7,10 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
 
 global $wpdb;
 
-$options = $wpdb->get_results( "SELECT option_name FROM ".$wpdb->options." WHERE option_name LIKE 'wpgdprc_%'");
-
-foreach( $options as $option ) {
-    delete_option( $option->option_name );
+// Remove everything related to the WP GDPR Compliance plugin
+$options = $wpdb->get_results("SELECT `option_name` FROM `" . $wpdb->options . "` WHERE `option_name` LIKE 'wpgdprc_%'");
+if ($options !== null) {
+    foreach ($options as $option) {
+        delete_option($option->option_name);
+    }
 }
